[WOR-802] Add ability to create temporary Azure Billing Projects  (#1296)
Context

We need the ability to create azure billing projects in Rawls for E2E testing.

This PR

Adds a withTemporaryAzureBillingProject test fixture that:
Accepts AzureManagedAppCoordinates and dispatches a call to Rawls via FirecloudOrch
Polls until the project reaches Ready status for 1 minutes; this is to accommodate the async nature of billing project and landing zone creation in Rawls.
Deletes the billing project and polls on queries for the project to return 404. We cannot fire and forget this call as the Azure billing project is not reusable until it is fully deleted (i.e,. the billing profile and landing zone have been deleted)
Note that GCP billing project creation now uses the polling codepaths and require that the project be "ready" before proceeding on with the test.